requirements.txt
alembic==1.5.8
altair==5.3.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
asyncgui==0.6.1
asynckivy==0.6.2
attrs==23.2.0
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.21.2
bleach==3.3.0
blinker==1.8.1
boto3==1.35.54
cachetools==5.3.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
crispy-bootstrap4==2022.1
cryptography==43.0.3
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1.2
django-allauth==65.3.0
django-autocomplete-light==3.11.0
django-crispy-forms==2.0
django-jalali==7.2.0
django-oauth-toolkit==3.0.1
django-querysetsequence==0.17
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
dnspython==2.1.0
docutils==0.20.1
dominate==2.6.0
email-validator==1.1.2
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-PageDown==0.3.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.0.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jwcrypto==1.5.6
Mako==1.1.4
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
materialyoucolor==2.0.9
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pillow==11.0.0
pipenv==2023.12.1
platformdirs==4.1.0
playwright==1.43.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.10
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.0.1
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-engineio==4.1.0
python-json-logger==2.0.7
python-socketio==5.2.1
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.3
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.5
sqlparse==0.5.1
stack-data==0.6.3
storage==0.0.4.3
streamlit==1.34.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.25.0
visitor==0.1.3
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
